+++ lib/decorators/types.ts
@@ -1,0 +1,5 @@
+export type DecoratorFunction = (
+  target: any,
+  propertyKey: string,
+  propertyDescriptor: PropertyDescriptor
+) => void;